requirements.txt
# Flask e extensões
flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Processamento de áudio e vídeo
openai-whisper==20231117
yt-dlp==2023.12.30
ffmpeg-python==0.2.0

# PyTorch
torch==2.1.0
torchaudio==2.1.0

# Interface e formatação
rich==13.7.0
colorama==0.4.6

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
tqdm==4.66.1

# Processamento e codificação
charset-normalizer==3.3.2
more-itertools==10.1.0
regex==2023.10.3